Finished application permission control

+++ routes/api/role.php
@@ -1,4 +1,5 @@
 <?php
+use Illuminate\Support\Facades\Route;
 
 /*
 |--------------------------------------------------------------------------
@@ -8,12 +9,22 @@
 */
 Route::group(["middleware" => "auth:api"], function () {
     Route::group(["middleware" => ["role:admin"]], function () {
-        Route::get("/roles", "API\RolesController@index");
-        Route::get("/role/{id}", "API\RolesController@show");
-        Route::delete("/role/{id}", "API\RolesController@destroy");
-        Route::post("/roles", "API\RolesController@store");
-        Route::put("/roles/{id}", "API\RolesController@update");
-        Route::patch("/roles/{roleId}/permission/{permissionId}", "API\RolesPermissionController@store");
-        Route::delete("/roles/{roleId}/permission/{permissionId}", "API\RolesPermissionController@destroy");
+        Route::get("/roles", "API\RolesController@index")->middleware("permission:read-roles");
+
+        Route::get("/role/{id}", "API\RolesController@show")->middleware("permission:read-roles");
+
+        Route::delete("/role/{id}", "API\RolesController@destroy")->middleware("permission:delete-roles");
+
+        Route::post("/roles", "API\RolesController@store")->middleware("permission:create-roles");
+
+        Route::put("/roles/{id}", "API\RolesController@update")->middleware("permission:update-roles");
+
+        Route::patch("/roles/{roleId}/permission/{permissionId}", "API\RolesPermissionController@store")->middleware(
+            "permission:update-roles",
+        );
+
+        Route::delete("/roles/{roleId}/permission/{permissionId}", "API\RolesPermissionController@destroy")->middleware(
+            "permission:delete-roles",
+        );
     });
 });
